+++ include/raytracer/utils.hpp
@@ -5,7 +5,7 @@
 
 constexpr double infinity = std::numeric_limits<double>::infinity();
 constexpr double pi = 3.1415926535897932385;
-constexpr float kEpsilon = 1e-8;
+constexpr double kEpsilon = 1e-8;
 
 constexpr double degrees_to_radians(double degrees) {
     return degrees * pi / 180.0;
